requirements.txt
beautifulsoup4==4.14.3
faiss-cpu==1.13.2
google-generativeai==0.5.4
httpx==0.28.1
huggingface_hub==1.10.2
langchain==1.2.15
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.29
langchain-experimental==0.4.1
langchain-google-genai==4.2.1
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langsmith==0.7.31
pathlib==1.0.1
pinecone==7.0.1
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
pydeck==0.9.1
pypdf==6.10.1
PyPDF2==3.0.1
python-dotenv==1.2.2
requests==2.33.1
sentence-transformers==5.4.1
streamlit==1.56.0